Rename ed25519hd.ts -> slip10.ts
