+++ packages/iov-keycontrol/types/keyring-entries/ed25519simpleaddress.d.ts
@@ -1,5 +1,5 @@
 import { KeyringEntryImplementationIdString, LocalIdentity } from "../keyring";
-import { Slip10KeyringEntry } from "./ed25519hd";
+import { Slip10KeyringEntry } from "./slip10";
 export declare class Ed25519SimpleAddressKeyringEntry extends Slip10KeyringEntry {
     static fromEntropy(bip39Entropy: Uint8Array): Ed25519SimpleAddressKeyringEntry;
     static fromMnemonic(mnemonicString: string): Ed25519SimpleAddressKeyringEntry;
