+++ packages/iov-keycontrol/types/keyring-entries/slip10.d.ts
